better List generation and function composition
Added overrides for list generation/comprehension so that the step in
between values (transform) and the equality checker can be supplied.
This is from a problem with using floats (where 1.0000001f doesn't equal
1).

Also fixed the definition of function composition as it was wrong. Now
it takes a function of A -> B and C -> A and returns one of C -> B.
Before it was doing A -> B, A-> B and returning A -> B. It was passing
tests though as it was doing Int -> Int and Int -> Int, which works out
correctly as it's essetntially A->A->A. however, when writing the demo
app I got a rather odd output by using int -> bool and int -> bool and
int -> bool. It compiled and outputted, thanks c++ (and my bad coding),
but was completely wrong.